Less templates, performance drop on GCC 5.2.1.
Performance drops because the comparison in index() is actually
performed, rather than compiled out. Loop unrolling would be
rather beneficial for the loop(s) around data::put, but I guess
one can't expect GCC to know that.

So I guess I'll be sticking with the templated version for now,
even though this one is nicer because of less (and more readable) code.

New measurements (GCC 5.2.1):
- base32enc file size MinSizeRel: 15136 bytes
- base32enc file size Release: 24104 bytes
- base32enc file size Default: 164896 bytes
- bench_base32_crockford MinSizeRel: ~1m20s
- bench_base32_crockford Release: 6.9-7s