Implement the foundations for base85/Z85.
This is not yet production ready because it doesn't currently
complain about input strings with incorrect lengths
(i.e. not matching the block size).

I'm committing it anyway because it's fully functional otherwise,
and reporting it as an error will involve changing the API in ways
that are maybe too large of a change by themselves to include in
another (this) commit.

In addition to implementing base85 decoding and the Z85 variant,
this commit also puts block decoding for codecs in a separate
struct named block_decoder. Both block_decoder and block_encoder
are now defined by the codec, which makes it possible to either
choose a helper struct (most codecs do for encoding) or perform
custom handling of the block, as is the case for base85.