Improve performance by checking for stop characters less often.
One of cppcodec's nicer features is correct handling of
string zero termination, in addition to padding, in decode().
This way, the caller can pass a fixed-size buffer containing
an unknown-size C string without having to perform strlen() on it.
The result buffer size will never grow larger than what's required
for decoding the string in its actual size. Frankly, this is neat
and some of the competition handles this in a much sloppier way.

There is one downside though, which is that we can't rely solely
on the given source size but additionally also have to check
whether the decoding loop should stop prematurely. A character
that causes us to exit the loop is called stop character here.

After previous optimization efforts, the check for stop characters
remained as the largest contributor to MSVC/x64 decoding times that
would negatively impact performance compared to the fastest other
base-N implementations.

This patch alleviates much of that pain by checking for
stop characters only once per decoded block. In order to
achieve this, we use the new representation of alphabet indexes
as 16 (or more) bit values with stop characters occupying an
exclusive bit range that isn't used for valid alphabet indexes.

Instead of checking for stop characters in each character
iteration, we now add up all the alphabet indexes via bitwise or
and continue doing so until the alphabet indexes for the block
have been gathered and it's time to actually decode them.
At that point, we can check for the presence of stop characters
with a single bitmasked-variable-is-nonzero check, and still exit
the loop on time before block decoding would otherwise write
extraneous characters to the result buffer.

A minor trade-off is that if a stop character is found, we need to
iterate through the last block once more to find out exactly where
it was, because the bitwise-or'd variable would not tell us that.
This adds a bit of extra code, but in practice is much faster
because it only has to be performed once per decoding run.

On top of that, this commit changes the newly simplified core loop
to use a pointer instead of an index to the current alphabet index
(representing an encoded character). This doesn't really add any
more complexity and further reduces the number of necessary
instructions, assuming the compiler didn't already optimize for it.

The difference on MSVC/x64 performance is striking. Here are some
size 32768 benchmarks for decoding on MinSizeRel builds
(measured on my laptop):

MSVC/x64 - string:   previously 329, now 87
MSVC/x64 - vector:   previously 327, now 86
MSVC/Win32 - string: previously 124, now 164
MSVC/Win32 - vector: previously  96, now 106

Curiously, this change degrades not just Win32 performance a bit
but also MSVC/x64 *encoding*, which goes up from 96 to 123
microseconds for a decoded-size 32768 buffer. I haven't changed
code related to encoding so I can only assume that the generated
binary is slightly larger and results in some extra cache misses.

TODO: Test on GCC & Clang